Add a check for consistency in string lengths, which determines what kind of col it is
